CLP: More clean-up related to serial port communication
- Change invesalius.data.trigger to invesalius.data.serial_port_connection

- Rename the class TriggerNew to SerialPortConnection

- Remove the old, unused Trigger class

- Rename a couple of variables in SerialPortConnection class

- Replace the incorrect docstring for SerialPortConnection class with a
  brief but more accurate docstring.
